Moved amm menu parts to a Menu class and changed sub-menu methods to private.
