+++ RebelAllianceBank/Classes/Currency.cs
@@ -2,6 +2,6 @@
 {
     public class Currency
     {
-
+        
     }
 }
